+++ src/Cooked/MockChain/Direct.hs
@@ -4,7 +4,6 @@
 -- internal state. This choice might be revised in the future.
 module Cooked.MockChain.Direct where
 
-import Cardano.Api qualified as Cardano
 import Cardano.Node.Emulator.Internal.Node qualified as Emulator
 import Control.Applicative
 import Control.Arrow
@@ -13,6 +12,7 @@ import Control.Monad.Except
 import Control.Monad.Identity
 import Control.Monad.Reader
 import Control.Monad.State.Strict
+import Control.Monad.Writer
 import Cooked.InitialDistribution
 import Cooked.MockChain.Balancing
 import Cooked.MockChain.BlockChain
@@ -22,16 +22,13 @@ import Cooked.MockChain.MockChainSt
 import Cooked.MockChain.UtxoState
 import Cooked.Output
 import Cooked.Skeleton
-import Data.Bifunctor (bimap)
 import Data.Default
-import Data.Map.Strict (Map)
 import Data.Map.Strict qualified as Map
 import Data.Set qualified as Set
 import Ledger.Index qualified as Ledger
 import Ledger.Orphans ()
 import Ledger.Tx qualified as Ledger
 import Ledger.Tx.CardanoAPI qualified as Ledger
-import PlutusLedgerApi.V3 qualified as Api
 
 -- * Direct Emulation
 
@@ -45,8 +42,8 @@ import PlutusLedgerApi.V3 qualified as Api
 -- 'Api.UtxoIndex', which we also keep in our state.
 
 newtype MockChainT m a = MockChainT
-  {unMockChain :: StateT MockChainSt (ExceptT MockChainError m) a}
-  deriving newtype (Functor, Applicative, MonadState MockChainSt, MonadError MockChainError)
+  {unMockChain :: (StateT MockChainSt (ExceptT MockChainError (WriterT [MockChainLogEntry] m))) a}
+  deriving newtype (Functor, Applicative, MonadState MockChainSt, MonadError MockChainError, MonadWriter [MockChainLogEntry])
 
 type MockChain = MockChainT Identity
 
@@ -59,10 +56,10 @@ instance (Monad m) => MonadFail (MockChainT m) where
   fail = throwError . FailWith
 
 instance MonadTrans MockChainT where
-  lift = MockChainT . lift . lift
+  lift = MockChainT . lift . lift . lift
 
 instance (Monad m, Alternative m) => Alternative (MockChainT m) where
-  empty = MockChainT $ StateT $ const $ ExceptT empty
+  empty = MockChainT $ StateT $ const $ ExceptT $ WriterT empty
   (<|>) = combineMockChainT (<|>)
 
 combineMockChainT ::
@@ -73,15 +70,17 @@ combineMockChainT ::
   MockChainT m x
 combineMockChainT f ma mb = MockChainT $
   StateT $ \s ->
-    let resA = runExceptT $ runStateT (unMockChain ma) s
-        resB = runExceptT $ runStateT (unMockChain mb) s
-     in ExceptT $ f resA resB
+    let resA = runWriterT $ runExceptT $ runStateT (unMockChain ma) s
+        resB = runWriterT $ runExceptT $ runStateT (unMockChain mb) s
+     in ExceptT $ WriterT $ f resA resB
+
+type MockChainReturn a b = (Either MockChainError (a, b), [MockChainLogEntry])
 
 mapMockChainT ::
-  (m (Either MockChainError (a, MockChainSt)) -> n (Either MockChainError (b, MockChainSt))) ->
+  (m (MockChainReturn a MockChainSt) -> n (MockChainReturn b MockChainSt)) ->
   MockChainT m a ->
   MockChainT n b
-mapMockChainT f = MockChainT . mapStateT (mapExceptT f) . unMockChain
+mapMockChainT f = MockChainT . mapStateT (mapExceptT (mapWriterT f)) . unMockChain
 
 -- | Executes a 'MockChainT' from some initial state; does /not/ convert the
 -- 'MockChainSt' into a 'UtxoState'.
@@ -89,8 +88,8 @@ runMockChainTRaw ::
   (Monad m) =>
   MockChainSt ->
   MockChainT m a ->
-  m (Either MockChainError (a, MockChainSt))
-runMockChainTRaw i0 = runExceptT . flip runStateT i0 . unMockChain
+  m (MockChainReturn a MockChainSt)
+runMockChainTRaw i0 = runWriterT . runExceptT . flip runStateT i0 . unMockChain
 
 -- | Executes a 'MockChainT' from an initial state set up with the given initial
 -- value distribution. Similar to 'runMockChainT', uses the default
@@ -100,45 +99,28 @@ runMockChainTFrom ::
   (Monad m) =>
   InitialDistribution ->
   MockChainT m a ->
-  m (Either MockChainError (a, UtxoState))
-runMockChainTFrom i0 = fmap (fmap $ second mcstToUtxoState) . runMockChainTRaw (mockChainSt0From i0)
+  m (MockChainReturn a UtxoState)
+runMockChainTFrom i0 s = first (right (second mcstToUtxoState)) <$> runMockChainTRaw (mockChainSt0From i0) s
 
 -- | Executes a 'MockChainT' from the canonical initial state and environment.
 -- The canonical environment uses the default 'SlotConfig' and
 -- @Cooked.Wallet.wallet 1@ as the sole wallet signing transactions.
-runMockChainT :: (Monad m) => MockChainT m a -> m (Either MockChainError (a, UtxoState))
+runMockChainT :: (Monad m) => MockChainT m a -> m (MockChainReturn a UtxoState)
 runMockChainT = runMockChainTFrom def
 
 -- | See 'runMockChainTRaw'
-runMockChainRaw :: MockChain a -> Either MockChainError (a, MockChainSt)
+runMockChainRaw :: MockChain a -> MockChainReturn a MockChainSt
 runMockChainRaw = runIdentity . runMockChainTRaw def
 
 -- | See 'runMockChainTFrom'
-runMockChainFrom :: InitialDistribution -> MockChain a -> Either MockChainError (a, UtxoState)
+runMockChainFrom :: InitialDistribution -> MockChain a -> MockChainReturn a UtxoState
 runMockChainFrom i0 = runIdentity . runMockChainTFrom i0
 
 -- | See 'runMockChainT'
-runMockChain :: MockChain a -> Either MockChainError (a, UtxoState)
+runMockChain :: MockChain a -> MockChainReturn a UtxoState
 runMockChain = runIdentity . runMockChainT
 
 -- * Direct Interpretation of Operations
-
-getIndex :: Ledger.UtxoIndex -> Map Api.TxOutRef Ledger.TxOut
-getIndex =
-  Map.fromList
-    . map (bimap Ledger.fromCardanoTxIn (Ledger.TxOut . toCtxTxTxOut))
-    . Map.toList
-    . Cardano.unUTxO
-  where
-    -- We need to convert a UTxO context TxOut to a Transaction context Tx out.
-    -- It's complicated because the datum type is indexed by the context.
-    toCtxTxTxOut :: Cardano.TxOut Cardano.CtxUTxO era -> Cardano.TxOut Cardano.CtxTx era
-    toCtxTxTxOut (Cardano.TxOut addr val d refS) =
-      let dat = case d of
-            Cardano.TxOutDatumNone -> Cardano.TxOutDatumNone
-            Cardano.TxOutDatumHash s h -> Cardano.TxOutDatumHash s h
-            Cardano.TxOutDatumInline s sd -> Cardano.TxOutDatumInline s sd
-       in Cardano.TxOut addr val dat refS
 
 instance (Monad m) => MonadBlockChainBalancing (MockChainT m) where
   getParams = gets mcstParams
@@ -146,6 +128,7 @@ instance (Monad m) => MonadBlockChainBalancing (MockChainT m) where
   txOutByRefLedger outref = gets $ Map.lookup outref . getIndex . mcstIndex
   datumFromHash datumHash = (txSkelOutUntypedDatum <=< Just . fst <=< Map.lookup datumHash) <$> gets mcstDatums
   utxosAtLedger addr = filter ((addr ==) . outputAddress . txOutV2FromLedger . snd) <$> allUtxosLedger
+  publish l = tell [l]
 
 instance (Monad m) => MonadBlockChainWithoutValidation (MockChainT m) where
   allUtxosLedger = gets $ Map.toList . getIndex . mcstIndex
@@ -155,6 +138,8 @@ instance (Monad m) => MonadBlockChainWithoutValidation (MockChainT m) where
 
 instance (Monad m) => MonadBlockChain (MockChainT m) where
   validateTxSkel skelUnbal = do
+    -- We log the submitted skeleton
+    gets mcstToSkelContext >>= publish . (`MCLogSubmittedTxSkel` skelUnbal)
     -- We retrieve the current parameters
     oldParams <- getParams
     -- We compute the optionally modified parameters
@@ -167,6 +152,8 @@ instance (Monad m) => MonadBlockChain (MockChainT m) where
     -- We balance the skeleton when requested in the skeleton option, and get
     -- the associated fee, collateral inputs and return collateral wallet
     (skel, fee, collateralIns, returnCollateralWallet) <- balanceTxSkel minAdaSkelUnbal
+    -- We log the adjusted skeleton
+    gets mcstToSkelContext >>= \ctx -> publish $ MCLogAdjustedTxSkel ctx skel fee collateralIns returnCollateralWallet
     -- We retrieve data that will be used in the transaction generation process:
     -- datums, validators and various kinds of inputs. This idea is to provide a
     -- rich-enough context for the transaction generation to succeed.
@@ -220,5 +207,7 @@ instance (Monad m) => MonadBlockChain (MockChainT m) where
       modify' (\st -> st {mcstCurrentSlot = mcstCurrentSlot st + 1})
     -- We return the parameters to their original state
     setParams oldParams
+    -- We log the validated transaction
+    publish $ MCLogNewTx (Ledger.fromCardanoTxId $ Ledger.getCardanoTxId cardanoTx)
     -- We return the validated transaction
     return cardanoTx