+++ tests/Cooked/MockChain/BlockChainSpec.hs
@@ -15,7 +15,7 @@ tests =
         "time handling"
         [ testProperty "bounds computed by slotToTimeInterval are included in slot" $
             \n ->
-              case runMockChain $ do
+              case fst $ runMockChain $ do
                 (l, r) <- slotToTimeInterval $ Ledger.Slot n
                 Ledger.Slot nl <- getEnclosingSlot l
                 Ledger.Slot nr <- getEnclosingSlot r
@@ -24,7 +24,7 @@ tests =
                 Right ((nl, nr), _) -> nl == n && nr == n,
           testProperty "bounds computed by slotToTimeInterval are maximal" $
             \n ->
-              case runMockChain $ do
+              case fst $ runMockChain $ do
                 (l, r) <- slotToTimeInterval $ Ledger.Slot n
                 Ledger.Slot nl <- getEnclosingSlot (l - 1)
                 Ledger.Slot nr <- getEnclosingSlot (r + 1)
@@ -32,7 +32,7 @@ tests =
                 Left _err -> False
                 Right ((nl, nr), _) -> nl == n - 1 && nr == n + 1,
           testProperty "time is always included in enclosing slot" $
-            \t -> case runMockChain $ slotToTimeInterval =<< getEnclosingSlot (Api.POSIXTime t) of
+            \t -> case fst $ runMockChain $ slotToTimeInterval =<< getEnclosingSlot (Api.POSIXTime t) of
               Left _err -> False
               Right ((Api.POSIXTime a, Api.POSIXTime b), _) -> a <= t && a <= b
         ]